Move PG manager to own utils file
