+++ python/ray/tune/ray_trial_executor.py
@@ -7,12 +7,12 @@ import random
 import time
 import traceback
 from contextlib import contextmanager
-from typing import Dict, List, Optional, Set
+from typing import List, Optional
 
 import ray
-from ray.actor import ActorClass, ActorHandle
+from ray.actor import ActorHandle
 from ray.exceptions import GetTimeoutError
-from ray import ObjectRef, ray_constants
+from ray import ray_constants
 from ray.resource_spec import ResourceSpec
 from ray.tune.durable_trainable import DurableTrainable
 from ray.tune.error import AbortTrialExecution, TuneError
@@ -20,6 +20,7 @@ from ray.tune.function_runner import FunctionRunner
 from ray.tune.logger import NoopLogger
 from ray.tune.result import TRIAL_INFO, STDOUT_FILE, STDERR_FILE
 from ray.tune.resources import PlacementGroupFactory, Resources
+from ray.tune.utils.placement_groups import PlacementGroupManager
 from ray.tune.utils.trainable import TrainableUtil
 from ray.tune.trial import Trial, Checkpoint, Location, TrialInfo
 from ray.tune.trial_executor import TrialExecutor
@@ -38,11 +39,6 @@ TUNE_RESULT_BUFFER_MIN_TIME_S = float(
     os.getenv("TUNE_RESULT_BUFFER_MIN_TIME_S", 0.))
 TUNE_RESULT_BUFFER_MAX_TIME_S = float(
     os.getenv("TUNE_RESULT_BUFFER_MAX_TIME_S", 100.))
-TUNE_MAX_PENDING_TRIALS_PG = int(os.getenv("TUNE_MAX_PENDING_TRIALS_PG", 1000))
-# Seconds we wait for a trial to come up before we make blocking calls
-# to process events
-TUNE_TRIAL_STARTUP_GRACE_PERIOD = float(
-    os.getenv("TUNE_TRIAL_STARTUP_GRACE_PERIOD", 10.))
 
 
 class _ActorClassCache:
@@ -145,207 +141,6 @@ class _TrialCleanup:
                 del self._cleanup_map[done]
 
 
-class _PlacementGroupManager:
-    def __init__(self):
-        self._staging: Dict[Trial, PlacementGroup] = {}
-        self._ready: Dict[Trial, PlacementGroup] = {}
-        self._staging_futures: Dict[ObjectRef, Trial] = {}
-        self._in_use: Set[PlacementGroup] = set()
-
-        self._latest_staging_start_time = time.time()
-
-    @property
-    def available_trials(self):
-        """Get trials with a ready placement group that have not been started.
-
-        Yields: Trial objects.
-
-        """
-        for ready, pg in self._ready.items():
-            if pg in self._in_use:
-                continue
-            yield ready
-
-    def stage_trial(self, trial: Trial):
-        """Stage a trial placement group.
-
-        Create the trial placement group if maximum number of pending
-        placement groups is not exhausted.
-
-        Args:
-            trial (Trial): Trial to stage.
-
-        Returns: False if trial has not been staged, True otherwise.
-
-        Creates its placement group and moves it to `self._staging`.
-        """
-        if trial in self._staging:
-            raise RuntimeError(f"Trial {trial} already staged.")
-
-        if not self.can_stage():
-            return False
-
-        pg = trial.placement_group_factory()
-
-        self._staging[trial] = pg
-        self._staging_futures[pg.ready()] = trial
-
-        self._latest_staging_start_time = time.time()
-
-        return True
-
-    def can_stage(self):
-        """Return True if we can stage another placement group."""
-        return len(self._staging) < TUNE_MAX_PENDING_TRIALS_PG
-
-    def update_status(self):
-        """Update placement group status.
-
-        Moves ready placement groups from `self._staging` to
-        `self._ready`.
-        """
-        ready = True
-        while ready:
-            # Use a loop as `ready` might return futures one by one
-            ready, _ = ray.wait(list(self._staging_futures.keys()), timeout=0)
-
-            for ready_fut in ready:
-                ready_trial = self._staging_futures.pop(ready_fut)
-                ready_pg = self._staging.pop(ready_trial)
-                self._ready[ready_trial] = ready_pg
-
-    def get_full_actor_cls(self, trial: Trial,
-                           actor_cls: ActorClass) -> Optional[ActorClass]:
-        """Get a fully configured actor class.
-
-        Returns the actor handle if the placement group is ready.
-
-        Args:
-            trial (Trial): trial object to start
-            actor_cls: Ray actor class.
-
-        Returns: Configured ActorClass or None
-
-        """
-        if trial not in self._ready:
-            return None
-
-        pg = self._ready[trial]
-        self._in_use.add(pg)
-
-        return actor_cls.options(
-            placement_group=pg, placement_group_bundle_index=0)
-
-    def get_trial_pg(self, trial: Trial) -> Optional[PlacementGroup]:
-        """Get placement group associated with trial.
-
-        Args:
-            trial (Trial): Trial object.
-
-        Returns: Placement group or None.
-
-        """
-        return self._ready.get(trial) or self._staging.get(trial)
-
-    def switch_trial_pgs(self, first: Trial, second: Trial):
-        """Switch placement groups associated with first and second trial.
-
-        Args:
-            first (Trial): Trial object.
-            second (Trial): Trial object.
-
-        """
-        # First, switch staging futures
-        # Note that zero, one, or both trials might be currently staging.
-        first_fut = second_fut = None
-        for future, trial in self._staging_futures.items():
-            if trial is first:
-                first_fut = future
-            if trial is second:
-                second_fut = future
-
-        if first_fut:
-            self._staging_futures[first_fut] = second
-
-        if second_fut:
-            self._staging_futures[second_fut] = first
-
-        # Then switch the references to the placement groups
-        if first in self._ready:
-            if second in self._ready:
-                first_pg = self._ready.pop(first)
-                second_pg = self._ready.pop(second)
-                self._ready[first] = second_pg
-                self._ready[second] = first_pg
-                return
-            elif second in self._staging:
-                first_pg = self._ready.pop(first)
-                second_pg = self._staging.pop(second)
-                self._staging[first] = second_pg
-                self._ready[second] = first_pg
-                return
-        elif first in self._staging:
-            if second in self._ready:
-                first_pg = self._staging.pop(first)
-                second_pg = self._ready.pop(second)
-                self._ready[first] = second_pg
-                self._staging[second] = first_pg
-                return
-            elif second in self._staging:
-                first_pg = self._staging.pop(first)
-                second_pg = self._staging.pop(second)
-                self._staging[first] = second_pg
-                self._staging[second] = first_pg
-                return
-        raise RuntimeError(
-            f"Could not switch placement groups of trial {first} and {second}")
-
-    def is_staging(self, trial: Trial) -> bool:
-        """Return True if trial placement group is staging.
-
-        Args:
-            trial (Trial): Trial object.
-
-        Returns: Boolean.
-
-        """
-        return trial in self._staging
-
-    def is_ready(self, trial: Trial) -> bool:
-        """Return True if trial placement group is ready.
-
-        Args:
-            trial (Trial): Trial object.
-
-        Returns: Boolean.
-
-        """
-        return trial in self._ready
-
-    def clean_trial_placement_group(self, trial: Trial):
-        """Remove reference to placement groups associated with a trial.
-
-        Args:
-            trial (Trial): Trial object.
-
-        Returns: None
-
-        """
-        pg = self._ready.pop(trial, None)
-        if pg and pg in self._in_use:
-            self._staging_futures.pop(pg, None)
-            self._in_use.remove(pg)
-
-        pg = self._staging.pop(trial, None)
-        if pg and pg in self._in_use:
-            self._staging_futures.pop(pg, None)
-            self._in_use.remove(pg)
-
-    def in_staging_grace_period(self):
-        return time.time(
-        ) <= self._latest_staging_start_time + TUNE_TRIAL_STARTUP_GRACE_PERIOD
-
-
 def noop_logger_creator(config, logdir):
     # Set the working dir in the remote process, for user file writes
     os.makedirs(logdir, exist_ok=True)
@@ -385,7 +180,7 @@ class RayTrialExecutor(TrialExecutor):
 
         self._avail_resources = Resources(cpu=0, gpu=0)
         self._committed_resources = Resources(cpu=0, gpu=0)
-        self._pg_manager = _PlacementGroupManager()
+        self._pg_manager = PlacementGroupManager()
 
         self._resources_initialized = False
